Relocate contrast buttons in Dock area
