+++ jungfrau_gui/ui_main_window.py
@@ -5,18 +5,24 @@ import pyqtgraph as pg
 from .ui_components.overlay import draw_overlay 
 from pyqtgraph.dockarea import Dock
 from PySide6.QtWidgets import (QMainWindow, QVBoxLayout, QWidget,
-                                QHBoxLayout, QPushButton,
-                                QMessageBox, QTabWidget)
+                                QHBoxLayout, QPushButton, QGridLayout,
+                                QMessageBox, QTabWidget, QLabel)
 from PySide6.QtCore import Qt, QObject, QEvent, QTimer
 from PySide6.QtGui import QShortcut, QKeySequence
 from .ui_components.visualization_panel.visualization_panel import VisualizationPanel
 from .ui_components.tem_controls.tem_controls import TemControls
 from .ui_components.file_operations.file_operations import FileOperations
 from .ui_components.utils import create_gaussian
+from .ui_components.toggle_button import ToggleButton
 
 import jungfrau_gui.ui_threading_helpers as thread_manager
 
 from importlib import resources
+
+from boost_histogram import Histogram
+from boost_histogram.axis import Regular
+
+from epoc import ConfigurationClient, auth_token, redis_host
 
 def get_git_info():
     # Load version from installed package resources
@@ -58,6 +64,8 @@ class ApplicationWindow(QMainWindow):
         self.initUI()
 
     def initUI(self):
+        self.cfg =  ConfigurationClient(redis_host(), token=auth_token())
+
         # Window Geometry
         self.setWindowTitle(self.version)
         self.setGeometry(50, 50, 1500, 1000)
@@ -100,7 +108,6 @@ class ApplicationWindow(QMainWindow):
 
         # Initial data (optional)
         data = create_gaussian(1000, globals.ncol, globals.nrow, 30, 15, np.deg2rad(35))
-        # data = np.random.rand(globals.nrow,globals.ncol).astype(globals.dtype)
         logging.debug(f"type(data) is {type(data[0,0])}")
         self.imageItem.setImage(data, autoRange = False, autoLevels = False, autoHistogramRange = False)
         
@@ -110,11 +117,55 @@ class ApplicationWindow(QMainWindow):
         # Mouse hovering
         self.imageItem.hoverEvent = self.imageHoverEvent
         
+        """ *********** """
+        """ Main Layout """
+        """ *********** """
+
         main_layout = QVBoxLayout()
         tools_layout = QHBoxLayout()
-        tools_layout.addWidget(self.dock,3)
 
-        # sections_layout = QHBoxLayout()
+        dock_layout = QVBoxLayout()
+
+        contrast_section = QVBoxLayout()
+
+        contrast_group = QGridLayout()
+        contrast_label = QLabel("Contrast Controls")
+        contrast_section.addWidget(contrast_label)
+
+        self.autoContrastBtn = ToggleButton('Apply Auto Contrast', self)
+        self.autoContrastBtn.setStyleSheet('background-color: green; color: white;')
+        self.autoContrastBtn.clicked.connect(self.toggle_autoContrast)
+        self.resetContrastBtn = QPushButton("Reset Contrast")
+        self.resetContrastBtn.clicked.connect(lambda: self.set_contrast(self.cfg.viewer_cmin, self.cfg.viewer_cmax))
+
+        self.contrast_0_Btn = QPushButton("-50 - 50")
+        self.contrast_1_Btn = QPushButton("0 - 100")
+        self.contrast_2_Btn = QPushButton("0 - 500")
+        self.contrast_3_Btn = QPushButton("0 - 1000")
+        self.contrast_4_Btn = QPushButton("0 - 1e5")
+
+        self.contrast_0_Btn.clicked.connect(lambda: self.set_contrast(-50, 50))
+        self.contrast_1_Btn.clicked.connect(lambda: self.set_contrast(0, 100))
+        self.contrast_2_Btn.clicked.connect(lambda: self.set_contrast(0, 500))
+        self.contrast_3_Btn.clicked.connect(lambda: self.set_contrast(0, 1000))
+        self.contrast_4_Btn.clicked.connect(lambda: self.set_contrast(0, 100000))
+
+        contrast_group.addWidget(self.autoContrastBtn, 0, 0,  1, 4)
+        contrast_group.addWidget(self.resetContrastBtn, 0, 4, 1, 4)
+
+        contrast_group.addWidget(self.contrast_0_Btn, 0, 8,  1, 1 )
+        contrast_group.addWidget(self.contrast_1_Btn, 0, 9,  1, 1 )
+        contrast_group.addWidget(self.contrast_2_Btn, 0, 10, 1, 1 )
+        contrast_group.addWidget(self.contrast_3_Btn, 0, 11, 1, 1 )
+        contrast_group.addWidget(self.contrast_4_Btn, 0, 12, 1, 1 )
+
+        contrast_section.addLayout(contrast_group)
+
+        dock_layout.addWidget(self.dock)
+        dock_layout.addLayout(contrast_section)
+
+        tools_layout.addLayout(dock_layout,3)
+
         tab_widget = QTabWidget()
 
         self.visualization_panel = VisualizationPanel(self)
@@ -124,7 +175,7 @@ class ApplicationWindow(QMainWindow):
         self.timer.timeout.connect(self.visualization_panel.captureImage)
 
         self.timer_contrast = QTimer(self)
-        self.timer_contrast.timeout.connect(self.visualization_panel.applyAutoContrast)
+        self.timer_contrast.timeout.connect(self.applyAutoContrast)
 
         self.tem_controls = TemControls(self)
         
@@ -156,14 +207,52 @@ class ApplicationWindow(QMainWindow):
         #self.view_box.addItem(self.imageItem)
         
         # Create a keyboard shortcut for Auto Range
-        #self.shortcut_autorange = QShortcut(QKeySequence("Ctrl+A"), self)
+        #self.shortcut_autorange = QShortcut(QKeySequence("A"), self)
         #self.shortcut_autorange.activated.connect(self.auto_range)
 
         logging.info("Viewer ready!")
 
     def auto_range(self):
         """Trigger the Auto Range (center image and scale view)."""
-        self.view_box.autoRange()
+        #self.view_box.autoRange()
+        self.plot.autoRange()
+
+    def set_contrast(self, lower, upper):
+        self.timer_contrast.stop()
+        self.autoContrastBtn.started = False
+        self.autoContrastBtn.setStyleSheet('background-color: green; color: white;')
+        self.autoContrastBtn.setText('Apply Auto Contrast')
+        self.histogram.setLevels(lower, upper)
+    
+    def toggle_autoContrast(self):
+        if not self.autoContrastBtn.started:
+            self.autoContrastBtn.setStyleSheet('background-color: red; color: white;')
+            self.autoContrastBtn.setText('Stop Auto Contrast')
+            self.autoContrastBtn.started = True
+            self.timer_contrast.start(10) # Assuming 100Hz streaming frequency at most
+        else:
+            self.timer_contrast.stop()
+            self.autoContrastBtn.started = False
+            self.autoContrastBtn.setStyleSheet('background-color: green; color: white;')
+            self.autoContrastBtn.setText('Apply Auto Contrast')
+    
+    # @profile
+    def applyAutoContrast(self, histo_boost = False):
+        if histo_boost:
+            data_flat = self.imageItem.image.flatten()
+            histogram = Histogram(Regular(1000000, data_flat.min(), data_flat.max()))
+            histogram.fill(data_flat)
+            cumsum_pre = np.cumsum(histogram.view())
+            cumsum = cumsum_pre[np.where(cumsum_pre < np.iinfo('int32').max-1)]
+            total = cumsum[-1]
+            low_thresh = np.searchsorted(cumsum, total * 0.01)
+            high_thresh = np.searchsorted(cumsum, total * 0.99999)
+        else:
+            image_data = self.imageItem.image
+            image_data_deloverflow = image_data[np.where(image_data < np.iinfo('int32').max-1)]
+            low_thresh, high_thresh = np.percentile(image_data_deloverflow, (1, 99.999))
+
+        self.histogram.setLevels(low_thresh, high_thresh)
 
     def roiChanged(self):
         roiPos = self.roi.pos()
